This is my instagram feed, incorporate the photos into the blogs. in addition give the photos names relevant to the design, a description etc... fitting the featured creations area
The top left one is `the defect` from slay the spire.
The top middle one is an archer which I designed.
The right top one a rooster.
The left bottom one is crested kingfisher.
The bottom middle one is surfing bird which I desgined.
Bottom right one is a duck.